Extract session middleware to a module
